requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
praw==7.7.0

# Celery + Redis (compatible versions)
celery[redis]==5.3.4
redis==4.6.0
django-redis==5.4.0

wikipedia-api==0.6.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
reportlab==4.0.7
weasyprint==60.1
django-ratelimit==4.1.0
pytest==7.4.3
pytest-django==4.7.0
